Imported Upstream version 2.17

+++ test/common.inc.sh
@@ -10,7 +10,7 @@ is_root () {
 }
 
 cleanup () {
-    rm -rf testtree
+    test -n "$TEST_NO_CLEANUP" && ! test "$TEST_NO_CLEANUP" = 0 || rm -rf testtree
 
     end
 }
@@ -25,5 +25,5 @@ prepare () {
     unset FAKECHROOT_CMD_SUBST FAKECHROOT_DEBUG FAKECHROOT_EXCLUDE_PATH
 }
 
-. "$srcdir/seq.inc"
-. "$srcdir/tap.inc"
+. "$srcdir/seq.inc.sh"
+. "$srcdir/tap.inc.sh"
